requirements.txt
numpy
pandas
scikit-learn
sqlalchemy
python-dotenv
nltk
pymysql
dvc


